TS-14 Make some InkAPI's 64-bit.
Also unify the APIs to use int32/uint32/int64/uint64.

Reviewed: John Plevyak

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1028930 13f79535-47bb-0310-9956-ffa450edef68

+++ example/append-transform/append-transform.c
@@ -330,7 +330,7 @@ load(const char *filename)
   INKFile fp;
   INKIOBufferBlock blk;
   char *p;
-  int avail;
+  int64 avail;
   int err;
 
   fp = INKfopen(filename, "r");
