+++ iocore/cache/I_Cache.h
@@ -183,7 +183,7 @@ struct CacheVConnection:public VConnection
   virtual int get_disk_io_priority() = 0;
   virtual bool set_pin_in_cache(time_t t) = 0;
   virtual time_t get_pin_in_cache() = 0;
-  virtual int get_object_size() = 0;
+  virtual int64 get_object_size() = 0;
 
   CacheVConnection();
 };
